+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## 1.0.0 - Unreleased
+## 1.1.0 - 2024-11-26
+### Added
+- Image field
+- Photo field
+
+## 1.0.0 - 2024-11-22
 ### Added
 - Initial release